+++ crates/node/lang/src/lib.rs
@@ -1,11 +1,11 @@
 pub mod node;
 pub mod npm;
-pub mod package;
+pub mod package_json;
 pub mod pnpm;
 pub mod yarn;
 
 pub use moon_lang::LockfileDependencyVersions;
-pub use package::PackageJson;
+pub use package_json::PackageJson;
 
 use moon_lang::{DependencyManager, Language, VersionManager};
