+++ crates/node/lang/src/package_json.rs
